+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.4.1 (2022-01-04)
+- Support random number generation for BigInt
+- Migrated project to Swift 5.5 and Xcode 13.2
+
 ## 2.4.0 (2021-05-12)
 - Several enhancements of the `Complex` type
 - Migrated project to Swift 5.4 and Xcode 12.5